(#7)controller 로직분리, route 딴 수정
